make sure clean doesn't delete important stuff

+++ clean.ps1
@@ -2,6 +2,9 @@
 . ".\harness-data.ps1"
 if ("Y" -eq (Read-Host "Enter Y to clear out all output files")) {
    foreach ($d in ($DATA.filePaths.GetEnumerator())) {
+      if (@("licenses","tests","diagnostics") -contains (Split-Path -Path $d.Key -Leaf)) {
+         continue
+      }
       if ((Test-Path $d.Value -PathType Container) -and "Y" -eq (Read-Host "Enter Y to nuke everything in $($d.Value)")) {
          Write-Host "Clearing .\$(Split-Path -Path $($d.Value) -Leaf)"
          Remove-Item -Recurse -Path "$($d.Value)\*"
